Lift course attr type interface out
