+++ src/components/SearchResult.vue
@@ -1,14 +1,9 @@
 <script setup lang="ts">
 import SearchBox from "./SearchBox.vue";
-import { Semester, CourseCategory } from "../common";
+import { CourseAttribute } from "../common";
 
 interface ResultEntry {
-  name: string;
-  id: string;
-  teacher: string;
-  grade: number;
-  semester: Semester;
-  category: CourseCategory;
+  attr: CourseAttribute;
   rate: number;
   lastRated: string;
 }
@@ -20,22 +15,26 @@ const getResultEntries = (
   console.log(`getting [${start}, ${end}) searching results...`);
   return [
     {
-      name: "计算机组成原理",
-      id: "XXX",
-      teacher: "ZK",
-      grade: 2.0,
-      semester: "春季学期",
-      category: "专业必修课",
+      attr: {
+        name: "计算机组成原理",
+        id: "XXX",
+        teacher: "ZK",
+        grade: 2.0,
+        semester: "春季学期",
+        category: "专业必修课",
+      },
       rate: 1.8,
       lastRated: "2021-11-05",
     },
     {
-      name: "原子物理",
-      id: "YYY",
-      teacher: "HHB",
-      grade: 3.0,
-      semester: "春季学期",
-      category: "专业必修课",
+      attr: {
+        name: "原子物理",
+        id: "YYY",
+        teacher: "HHB",
+        grade: 3.0,
+        semester: "春季学期",
+        category: "专业必修课",
+      },
       rate: 2.7,
       lastRated: "2020-7-05",
     },
@@ -55,19 +54,19 @@ const getResultEntries = (
       class="bg-gray-100 rounded-lg"
     >
       <div class="flex space-x-2 text-lg">
-        <div>{{ resultEntry.name }}</div>
-        <div>{{ resultEntry.id }}</div>
+        <div>{{ resultEntry.attr.name }}</div>
+        <div>{{ resultEntry.attr.id }}</div>
         <div>|</div>
         <div>主讲教师</div>
-        <div>{{ resultEntry.teacher }}</div>
+        <div>{{ resultEntry.attr.teacher }}</div>
         <div>|</div>
         <div>学分</div>
-        <div>{{ resultEntry.grade }}</div>
+        <div>{{ resultEntry.attr.grade }}</div>
         <div>|</div>
         <div>开课学期</div>
-        <div>{{ resultEntry.semester }}</div>
+        <div>{{ resultEntry.attr.semester }}</div>
         <div>|</div>
-        <div>{{ resultEntry.category }}</div>
+        <div>{{ resultEntry.attr.category }}</div>
       </div>
       <div class="mt-2 indent-4 text-lg">
         综合评分 {{ resultEntry.rate }}，最近一条点评为
